requirements.txt
grpcio==1.17.0
grpcio-tools==1.17.0
flake8==3.6.0